Add `move_to_list` to `task::Repo`
Introduces the `move_to_list` method in `Repo` for updating a task's
associated list. This leverages the `list::Name` enum for stricter type
safety and ensures cleaner interactions with the database.

Adds a SQL constant `SET_LIST_ID_SQL` for the update query and a unit
test to verify correct functionality.

+++ crates/complish/src/task/mod.rs
@@ -19,7 +19,7 @@ pub struct Task {
   #[get = "pub"]
   id: u32,
   #[get = "pub"]
-  list_id: u32,
+  list_id: u8,
   #[get = "pub"]
   status: Status,
   #[get = "pub"]
